# rest-api/pnc-openapi.yaml
openapi: 3.0.1
info:
  title: PNC
  description: PNC build system
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 3.3.2-SNAPSHOT
servers:
- url: /pnc-rest
  variables: {}
paths:
  /artifacts:
    get:
      tags:
      - Artifacts
      summary: Gets all artifacts.
      operationId: getAll
      parameters:
      - name: pageIndex
        in: query
        description: Index of the page to return. Index starts with 0.
        schema:
          type: integer
          format: int32
          default: 0
      - name: pageSize
        in: query
        description: Number of entries that should be included in a page. Maximum
          page size is 200.
        schema:
          type: integer
          format: int32
          default: 50
      - name: sort
        in: query
        description: "Sorting RSQL. Format: sort=asc=path.to.field (or =desc=)."
        schema:
          type: string
      - name: q
        in: query
        description: RSQL Query.
        schema:
          type: string
      - name: sha256
        in: query
        description: Filter by sha256 of the artifact.
        schema:
          type: string
      - name: md5
        in: query
        description: Filter by md5 of the artifact.
        schema:
          type: string
      - name: sha1
        in: query
        description: Filter by sha1 of the artifact.
        schema:
          type: string
      responses:
        "200":
          description: Success with results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactPage'
        "400":
          description: Invalid input parameters or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - Internal
      - Artifacts
      summary: "[role:pnc-app-artifact-user, pnc-users-admin] Creates a new Artifact."
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Artifact'
        required: true
      responses:
        "201":
          description: Entity successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artifact'
        "400":
          description: Invalid input parameters or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Conflict while saving an entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /artifacts/filter:
    get:
      tags:
      - Artifacts
      summary: Gets all artifacts according to specified filters.
      operationId: getAllFiltered
      parameters:
      - name: pageIndex
        in: query
        description: Index of the page to return. Index starts with 0.
        schema:
          type: integer
          format: int32
          default: 0
      - name: pageSize
        in: query
        description: Number of entries that should be included in a page. Maximum
          page size is 200.
        schema:
          type: integer
          format: int32
          default: 50
      - name: identifier
        in: query
        description: Filter by artifact identifier or its part.
        schema:
          type: string
      - name: qualities
        in: query
        description: List of artifact qualities to include in result.
        schema:
          uniqueItems: true
          type: array
          items:
            type: string
            enum:
            - NEW
            - VERIFIED
            - TESTED
            - DEPRECATED
            - BLACKLISTED
            - DELETED
            - TEMPORARY
            - IMPORTED
      - name: repoType
        in: query
        description: Type of target repository.
        schema:
          type: string
          enum:
          - MAVEN
          - NPM
          - COCOA_POD
          - GENERIC_PROXY
          - DISTRIBUTION_ARCHIVE
      - name: buildCategories
        in: query
        description: List of artifact build categories to include in result.
        schema:
          uniqueItems: true
          type: array
          items:
            type: string
            enum:
            - STANDARD
            - SERVICE
            - AUTO
      responses:
        "200":
          description: Success with results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactInfoPage'
        "400":
          description: Invalid input parameters or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /artifacts/purl/{purl}:
    get:
      tags:
      - Artifacts
      summary: Gets a specific artifact by purl of the artifact.
      operationId: getSpecificFromPurl
      parameters:
      - name: purl
        in: path
        description: Purl of the artifact
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Success with results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artifact'
        "404":
          description: Can not find specified result
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /artifacts/{id}:
    get:
      tags:
      - Artifacts
      summary: Gets a specific artifact.
      operationId: getSpecific
      parameters:
      - name: id
        in: path
        description: ID of the artifact
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Success with results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artifact'
        "404":
          description: Can not find specified result
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
      - Internal
      - Artifacts
      summary: "[role:pnc-app-artifact-user, pnc-users-admin] Updates an existing\
        \ Artifact"
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Artifact'
        required: true
      responses:
        "204":
          description: Entity successfully updated
        "400":
          description: Invalid input parameters or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Conflict while saving an entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /artifacts/{id}/artifacts/quality:
    post:
      tags:
      - Artifacts
      summary: "Add a new quality level revision for this artifact. Accepted values\
        \ from standard users are NEW, VERIFIED, TESTED, DEPRECATED. Users with pnc-app-artifact-user,\
        \ pnc-app-build-user, pnc-users-admin role can also specify BLACKLISTED and\
        \ DELETED quality levels."
      operationId: createQualityLevelRevision
      parameters:
      - name: id
        in: path
        description: ID of the artifact
        required: true
        schema:
          type: string
      - name: quality
        in: query
        description: Quality level of the artifact.
        required: true
        schema:
          type: string
      - name: reason
        in: query
        description: The reason for adding a new quality level for this artifact.
        required: true
        schema:
          type: string
      responses:
        "201":
          description: Entity successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactRevision'
        "400":
          description: Invalid input parameters or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Conflict while saving an entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /artifacts/{id}/dependant-builds:
    get:
      tags:
      - Artifacts
      summary: Gets the build(s) that depends on this artifact.
      operationId: getDependantBuilds
      parameters:
      - name: id
        in: path
        description: ID of the artifact
        required: true
        schema:
          type: string
      - name: pageIndex
        in: query
        description: Index of the page to return. Index starts with 0.
        schema:
          type: integer
          format: int32
          default: 0
      - name: pageSize
        in: query
        description: Number of entries that should be included in a page. Maximum
          page size is 200.
        schema:
          type: integer
          format: int32
          default: 50
      - name: sort
        in: query
        description: "Sorting RSQL. Format: sort=asc=path.to.field (or =desc=)."
        schema:
          type: string
      - name: q
        in: query
        description: RSQL Query.
        schema:
          type: string
      - name: latest
        in: query
        description: Should return only latest build?
        schema:
          type: boolean
          default: false
      - name: running
        in: query
        description: Should return only running builds?
        schema:
          type: boolean
          default: false
      - name: buildConfigName
        in: query
        description: Filters builds by BuildConfig name in a revision used to trigger
          the build. Supports LIKE queries in style *name-part*
        schema:
          type: string
          default: ""
      responses:
        "200":
          description: Success with results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildPage'
        "400":
          description: Invalid input parameters or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /artifacts/{id}/revisions:
    get:
      tags:
      - Artifacts
      summary: Gets audited revisions of this artifact.
      operationId: getRevisions
      parameters:
      - name: id
        in: path
        description: ID of the artifact
        required: true
        schema:
          type: string
      - name: pageIndex
        in: query
        description: Index of the page to return. Index starts with 0.
        schema:
          type: integer
          format: int32
          default: 0
      - name: pageSize
        in: query
        description: Number of entries that should be included in a page. Maximum
          page size is 200.
        schema:
          type: integer
          format: int32
          default: 50
      responses:
        "200":
          description: Success with results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactRevisionPage'
        "400":
          description: Invalid input parameters or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ArtifactInfoPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ArtifactInfo'
        pageIndex:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalHits:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
    ArtifactPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Artifact'
        pageIndex:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalHits:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
    ArtifactRevision:
      required:
      - id
      type: object
      properties:
        artifactQuality:
          type: string
          enum:
          - NEW
          - VERIFIED
          - TESTED
          - DEPRECATED
          - BLACKLISTED
          - DELETED
          - TEMPORARY
          - IMPORTED
        buildCategory:
          type: string
          enum:
          - STANDARD
          - SERVICE
          - AUTO
        id:
          type: string
        modificationTime:
          type: string
          format: date-time
        modificationUser:
          $ref: '#/components/schemas/User'
        qualityLevelReason:
          type: string
        rev:
          type: integer
          format: int32
...
    ErrorResponse:
      type: object
      properties:
        details:
          type: object
        errorMessage:
          type: string
        errorType:
          type: string
...